% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectMsSql.R
\name{connectMsSqlDomainLogin}
\alias{connectMsSqlDomainLogin}
\title{Connects an R session to a running instance of Microsoft SQL Server 
using Active Directory domain login credentials over ODBC.  When using Kerberos or
Windows Integrated Authentication (SSPI) credentials, select the "ODBC Driver 17 for SQL Server"
ODBC driver.  When it is necessarry to send the user name and password as function arguments,
select the "FreeTDS" driver.  The former method is more secure and should be preferred when
possible.  The two drivers are mutually exclusive in their ability to authenticate domain
accounts in these two fashions.}
\usage{
connectMsSqlDomainLogin(server, user = NA, domain = NA,
  password = NA, database, port = 1433,
  driver = c("ODBC Driver 17 for SQL Server", "FreeTDS"),
  tdsVersion = "8.0")
}
\arguments{
\item{server}{The network address of the Windows instance that is running SQL Server.}

\item{user}{The SQL Server user name.}

\item{domain}{The Active Directory domain for the login.}

\item{password}{The password for the SQl Server user.}

\item{database}{The name of the database on the server that you want to connect to.}

\item{port}{The TCP port that SQL Server is listening on.}

\item{driver}{The name of an installed ODBC driver to use for the connection.  Current supported options are "FreeTDS" and "ODBC Driver 17 for SQL Server"}

\item{tdsVersion}{The TDS protocol version to use for the connection.}
}
\description{
Connects an R session to a running instance of Microsoft SQL Server 
using Active Directory domain login credentials over ODBC.  When using Kerberos or
Windows Integrated Authentication (SSPI) credentials, select the "ODBC Driver 17 for SQL Server"
ODBC driver.  When it is necessarry to send the user name and password as function arguments,
select the "FreeTDS" driver.  The former method is more secure and should be preferred when
possible.  The two drivers are mutually exclusive in their ability to authenticate domain
accounts in these two fashions.
}
\keyword{ODBC}
\keyword{SQL}
\keyword{Server}
